+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityTableModel.java
@@ -22,6 +22,7 @@ import is.codion.common.Conjunction;
 import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.model.table.ColumnConditionModel;
 import is.codion.common.model.table.TableConditionModel;
+import is.codion.common.observer.Observable;
 import is.codion.common.observer.Observer;
 import is.codion.common.state.State;
 import is.codion.common.state.StateObserver;
@@ -348,7 +349,7 @@ public class SwingEntityTableModel implements EntityTableModel<SwingEntityEditMo
 	@Override
 	public final Collection<Entity> find(Collection<Entity.Key> keys) {
 		requireNonNull(keys, "keys");
-		return items().stream()
+		return items().get().stream()
 						.filter(entity -> keys.contains(entity.primaryKey()))
 						.collect(toList());
 	}
@@ -379,7 +380,7 @@ public class SwingEntityTableModel implements EntityTableModel<SwingEntityEditMo
 	}
 
 	@Override
-	public final Collection<Entity> items() {
+	public Observable<Collection<Entity>> items() {
 		return tableModel.items();
 	}